fix(dbus): hold tile batches behind the pending screens announce
Focusing the middle of three strip columns from another desktop left
it at its old rect while its neighbour slid away. The daemon's tile
batch reached the effect before the coalesced desktop-switch announce.
The effect applied the first entry synchronously, then the announce
bumped its global stagger epoch and dropped the remaining staggered
entries as superseded. The scroll engine's emit-on-change baseline
already believed the batch landed and the force-emit arm had been
spent on it, so the two queued retiles that followed were suppressed
and nothing ever re-sent the rects.

TilingAdaptor::relayTileRequestsJson now holds any batch that arrives
while the screens announce is queued and the announce lambda flushes
them in arrival order right after it emits, ahead of the parked-open
retry. clearEngine sweeps held batches with the announce they waited
on. Both engine emit sites only fire for screens the engine owns
post-flip, so a held batch never describes a removed screen.

The engine relay entry points move to src/dbus/tilingadaptor/relay.cpp
so the grandfathered adaptor file does not grow past the ceiling.

Regression test: testTileBatchHeldBehindPendingScreensAnnounce, which
fails without the hold.
